requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coverage==4.5.3
cryptography==2.5
defusedxml==0.5.0
Django==2.1.5
django-markup==1.3
djangorestframework==3.9.1
html5lib==1.0.1
idna==2.8
isodate==0.6.0
lxml==4.3.3
mock==2.0.0
mysqlclient==1.3.14
pbr==5.1.1
pkgconfig==1.5.1
pycparser==2.19
pyOpenSSL==19.0.0
python3-saml==1.6.0
pytz==2018.9
requests==2.22.0
six==1.12.0
textile==3.0.3
urllib3==1.25.7
uWSGI==2.0.17.1
webencodings==0.5.1
xmlsec==1.3.3
